Move shared static functions to utils
This allows other modules like the Celery worker to use them.
